requirements.txt
python-slugify[unicode]
pymysql